🪛[chore]: 커스텀 드롭다운 UI 수정
